Fix linters error 7.0
